replace SignInButton with GoogleSignIn; in SignUp, move createUser inside function
